Add Suspense support for home page
